double type

+++ headers/Stack/Stack.h
@@ -4,7 +4,7 @@
 #ifndef STACK_H_
 #define STACK_H_
 
-typedef unsigned long long elem_t;
+typedef double elem_t;
 
 const ssize_t ReallocationScale = 2;
 const ssize_t InitialCapacity = 10;
